+++ release_notes.md
@@ -1,5 +1,15 @@
 # Release Notes - blacklight-cornell-requests
 
+## v1.2.2
+
+### Enhancements
+- Cleaned up markup in request forms
+- Updated tests
+- Added item location to copy selection view (DISCOVERYACCESS-2278)
+
+### Bug fixes
+
+
 ## v1.2.1
 - Improved parsing of ISBNs for Borrow Direct searches